Added features to handle copy, move, rename errors; updated file scan query, user menu
Added testing after each cp and mv, to ensure that program skips file or stops if cp or mv should fail (in particular, a mv rename won't even be attempted if the previous cp has failed).

Improved scan to detect the prefix "attendancelog" when scanning for files to report or rename. This allows the user to rename files to all available options (before files prefixed with "attendancelog" did not show up in the scan).

Updated all menus to allow for help and usage information  from the standard user menu, there is no longer a need to just "know" to add -h as an option.